+++ app/src/main/java/org/fuusio/kide/app/feature/about/AboutFeature.kt
@@ -17,14 +17,20 @@
 package org.fuusio.kide.app.feature.about
 
 import android.content.Context
+import org.fuusio.kide.app.feature.about.navigation.AboutNavKey
 import org.fuusio.kide.app.feature.about.presentation.AboutProcessor
 import org.fuusio.kide.app.feature.about.presentation.AboutViewState
 import org.fuusio.kide.app.FuusioApplicationFeature
 import org.fuusio.kide.feature.KoinFeature
+import org.fuusio.kide.navigation.ScreenNavKeyRegistry
 import org.koin.dsl.module
 
 object AboutFeature : KoinFeature {
-    
+
+    override fun initialize() {
+        ScreenNavKeyRegistry.register(AboutNavKey)
+    }
+
     override fun koinModule(applicationContext: Context) = module {
 
         includes(FuusioApplicationFeature.koinModule(applicationContext))